+++ internal/mdathome/constants.go
@@ -2,7 +2,7 @@ package mdathome
 
 // Client settings
 const clientVersion string = "v1.5.2"
-const specVersion int = 18
+const specVersion int = 19
 
 // Backend settings (Swap to use testnet)
 const apiBackend string = "https://api.mangadex.network" // "https://mangadex-test.net"\