+++ CHANGELOG.md
@@ -6,11 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- [2020-09-08] Added client configuration of optional HTTP2 by [@lflare].
-- [2020-08-23] Added some form of image integrity check via use of SHA256 checksums provided by upstream by [@lflare].
 
 ### Changed
-- [2020-09-08] Bumped client version up to 19 by [@lflare].
 
 ### Deprecated
 
@@ -19,6 +16,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v1.5.2] - 2020-09-08
+### Added
+- [2020-09-08] Added client configuration of optional HTTP2 by [@lflare].
+- [2020-08-23] Added some form of image integrity check via use of SHA256 checksums provided by upstream by [@lflare].
+
+### Changed
+- [2020-09-08] Bumped client version up to 19 by [@lflare].
 
 ## [v1.5.1] - 2020-08-16
 ### Fixed